feat(authentification): working tokens and validations

+++ vaccinationapp/app/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render
 from rest_framework import generics, status, viewsets
-from .serializers import RegisterSerializer, UserActivationTokenSerializer, LoginSerializer, CountySerializer, CitySerializer, VaccineSerializer, CategorySerializer, OfficeSerializer, PersonSerializer, AppointmentSerializer, WaitingSerializer
+from .serializers import RegisterSerializer, LoginSerializer, CountySerializer, CitySerializer, VaccineSerializer, CategorySerializer, OfficeSerializer, PersonSerializer, AppointmentSerializer, WaitingSerializer, ResetPasswordEmailRequestSerializer
 from rest_framework.response import Response
 from rest_framework_simplejwt.tokens import RefreshToken
 from .models import User, UserActivationToken, County, City, Vaccine, Categories, Office, Person, Appointment, Waiting
@@ -8,11 +8,15 @@ from .utils import Util
 from django.contrib.sites.shortcuts import get_current_site
 from django.urls import reverse
 from django.shortcuts import get_object_or_404
-import datetime
+import jwt, datetime
 from rest_framework.views import APIView
+from django.urls import reverse
+from rest_framework.permissions import IsAuthenticated
 
 
 class RegisterView(generics.GenericAPIView):
+  permission_classes = (IsAuthenticated,)
+
   serializer_class = RegisterSerializer
 
   def post(self, request):
@@ -24,8 +28,9 @@ class RegisterView(generics.GenericAPIView):
     user = User.objects.get(email=user_data['email'])
     
     token = RefreshToken.for_user(user).access_token
+    
     current_site = get_current_site(request).domain
-
+    # relativeLink = reverse('verify')
     now = datetime.datetime.now()
     UserToken = UserActivationToken()
    
@@ -44,14 +49,20 @@ class Verify(generics.GenericAPIView):
   serializer_class = UserActivationToken
 
   def get(self,request,token):
-    userToken = UserActivationToken.objects.get(token=token)
-    user = userToken.getUser()
-    userDb = User.objects.get(id = user)
-    userDb.is_active = True
-    userDb.save()
-    userToken.delete()
-    return Response({'email': 'Successfully activated'}, status=status.HTTP_200_OK)
-
+    try:
+      userToken = UserActivationToken.objects.get(token=token)
+      user = userToken.getUser()
+      userDb = User.objects.get(id = user)
+      if not userDb.is_active:
+        userDb.is_active = True
+        userDb.save()
+        userToken.delete()
+      return Response({'email': 'Successfully activated'}, status=status.HTTP_200_OK)
+    except jwt.ExpiredSignatureError as identifier:
+      return Response({'error': 'Activation Expired'}, status=status.HTTP_400_BAD_REQUEST)
+    except jwt.exceptions.DecodeError as identifier:
+      return Response({'error': 'Invalid token'}, status=status.HTTP_400_BAD_REQUEST)
+  
 
 
 class Login(generics.GenericAPIView):
@@ -62,6 +73,22 @@ class Login(generics.GenericAPIView):
     serializer.is_valid(raise_exception=True)
 
     return Response(serializer.data, status=status.HTTP_200_OK)
+
+
+
+class RequestPasswordResetEmail(generics.GenericAPIView):
+    serializer_class = ResetPasswordEmailRequestSerializer
+
+    def post(self, request):
+      data={'request':request, 'data':request.data}
+      serializer = self.serializer_class(data=request.data)
+      serializer.is_valid(raise_exception=True)
+      return Response({'success':'We have sent you a link to reset your password.'}, status=status.HTTP_200_OK)
+
+
+class PassordTokenCheck(generics.GenericAPIView):
+  def get(self,request, uidb64,token):
+    pass
